+++ tests/test_connect.py
@@ -38,3 +38,23 @@ class SocketErrorTestCase(BaseTestCase):
 
             with self.assertRaises(errors.SocketTimeoutError):
                 self.client.execute('SHOW TABLES')
+
+    def test_transport_not_connection_on_disconnect(self):
+        # Create connection.
+        self.client.execute('SELECT 1')
+
+        def side_effect(*args, **kwargs):
+            # Exception should be caught in graceful disconnect.
+            raise socket.error(107, 'Transport endpoint is not connected')
+
+        connection = self.client.connection
+
+        with patch.object(connection, 'ping') as mocked_ping:
+            mocked_ping.return_value = False
+
+            with patch.object(connection, 'socket') as mocked_socket:
+                mocked_socket.shutdown.side_effect = side_effect
+
+                # New socket should be created.
+                rv = self.client.execute('SELECT 1')
+                self.assertEqual(rv, [(1, )])